pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelarter"
version = "0.1.0"
description = "A localized restoration and discretization tool for nearly-pixel-art images."
readme = "README.md"
authors = [
    { name = "dok-the-blogger" }
]
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
]

[tool.ruff.lint.isort]
known-first-party = ["pixelarter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
